+++ andes/main.py
@@ -676,7 +676,7 @@ def prepare(quick=False, incremental=False, cli=False, full=False, **kwargs):
     if full is True:
         quick = False
 
-    system = System()
+    system = System(options=kwargs)
     system.prepare(quick=quick, incremental=incremental)
 
     if cli is True:
@@ -723,17 +723,13 @@ def selftest(quick=False, **kwargs):
     sys.stdout = sys.__stdout__
 
 
-def doc(attribute=None, list_supported=False, init_seq=False, config=False, **kwargs):
+def doc(attribute=None, list_supported=False, config=False, **kwargs):
     """
     Quick documentation from command-line.
     """
     system = System()
     if attribute is not None:
         if attribute in system.__dict__ and hasattr(system.__dict__[attribute], 'doc'):
-            if init_seq is True:
-                system.__dict__[attribute].get_init_order()
-                return
-
             logger.info(system.__dict__[attribute].doc())
         else:
             logger.error('Model <%s> does not exist.', attribute)